[refactor] #252 refactor timer view
